+++ analyze_facebook_data.rb
@@ -9,6 +9,7 @@ require_relative 'lib/analyze_facebook_data.rb'
 require_relative 'lib/making_friends_data.rb'
 require_relative 'lib/messages_sent.rb'
 require_relative 'lib/friends_dates.rb'
+require_relative 'lib/contact_list.rb'
 
 # images exchanged
 # links exchanged
@@ -146,40 +147,6 @@ package.workbook.add_worksheet(name: 'Vocabulary statistics') do |sheet|
   words_ranked.each do |word, count|
     sheet.add_row [rank, word, count]
     rank += 1
-  end
-end
-
-class ContactList
-  # contact info data
-  # how much facebook archived
-  attr_reader :contacts, :catalog
-
-  def initialize(data_catalog)
-    @catalog = data_catalog
-  end
-
-  def run
-    @contacts = contacts_rows.map do |contact|
-      text = contact.text
-      if text == 'NameContacts'
-      else
-        contact_info = text.split('contact: ')
-        [String(contact_info[0]), contact_info[1..3].join(' ')]
-      end
-    end.compact.uniq
-    self
-  end
-
-  def doc
-    Nokogiri::HTML(content)
-  end
-
-  def content
-    File.open("#{catalog}/html/contact_info.htm").read
-  end
-
-  def contacts_rows
-    doc.css('div.contents tr')
   end
 end
